backend/requirements.txt
langchain-openai
python-dotenv
openai
Flask
Flask-MySQLdb
Werkzeug
Flask-Cors